requirements.txt

tensorflow
keras
Flask
numpy
Pillow
pandas
matplotlib
seaborn
